pyproject.toml
[project]
name = "vic-worksheets"
version = "0.1.0"
description = "Homework resources"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-generativeai>=0.8.5",
    "jupyter>=1.1.1",
    "matplotlib>=3.10.3",
    "numpy>=2.3.2",
    "python-docx>=1.1.2",
    "python-dotenv>=1.1.0",
]
